All ES6 all the time
Clean up tests
Minify es6 bundle with babili
Convert custom-style to es6 class
Use WCT@next for magic server-side compilation
