+++ index.js
@@ -7,9 +7,11 @@ The complete set of contributors may be found at http://polymer.github.io/CONTRI
 Code distributed by Google as part of the polymer project is also
 subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
 */
+'use strict';
 
 /*
-A simple shim to generate a testable module
+Small module to load ShadyCSS and CustomStyle together
 */
 
-import '../../src/custom-style'
+import './src/ShadyCSS'
+import './src/custom-style-element'
