+++ package.json
@@ -21,20 +21,15 @@
   },
   "homepage": "http://webcomponents.org",
   "devDependencies": {
-    "@webcomponents/custom-elements": "^1.0.0-alpha.3",
-    "babel-plugin-external-helpers": "^6.8.0",
-    "babel-preset-es2015": "^6.16.0",
+    "babel-preset-babili": "0.0.9",
     "del": "^2.2.2",
     "eslint": "^3.4.0",
     "eslint-plugin-html": "^1.5.2",
-    "google-closure-compiler": "^20160822.2.0",
     "gulp": "^3.8.8",
+    "gulp-babel": "^6.1.2",
     "gulp-rename": "^1.2.2",
-    "gulp-sourcemaps": "^1.6.0",
-    "rollup-plugin-babel": "^2.6.1",
-    "rollup-stream": "^1.13.0",
-    "vinyl-buffer": "^1.0.0",
-    "vinyl-source-stream": "^1.1.0",
-    "web-component-tester": "^4.0.1"
+    "gulp-rollup": "^2.5.1",
+    "gulp-sourcemaps": "^1.9.1",
+    "web-component-tester": "^6.0.0-prerelease.2"
   }
 }
