+++ tests/css-parse.html
@@ -14,7 +14,7 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
   <meta charset="utf-8">
 
   <script src="../../web-component-tester/browser.js"></script>
-  <script src="../../webcomponentsjs/webcomponents-lite.js"></script>
+  <!--<script src="../../webcomponentsjs/webcomponents-lite.js"></script>-->
   <script src="module/generated/css-parse.js"></script>
 
   <title>css-parse</title>
@@ -104,15 +104,14 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
   }
 
   suite('css-parse', function() {
-    var css, s, tree;
+    var s, tree;
 
     setup(function() {
-      css = window.CssParse;
       s = document.querySelector('style#test');
-      tree = css.parse(s.textContent);
+      tree = window.CssParse.parse(s.textContent);
     });
 
-    test('css rules parse', function() {
+    test('window.CssParse rules parse', function() {
       assert.equal(tree.rules.length, 4, 'unexpected number of rules');
       assert.equal(tree.rules[2].rules.length, 8, 'unexpected number of rules in keyframes');
       assert.equal(tree.rules[3].rules.length, 1, 'unexpected number of rules in @media');
@@ -133,32 +132,32 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
     });
 
     test('rule types', function() {
-      assert.equal(tree.rules[0].type, css.types.STYLE_RULE);
-      assert.equal(tree.rules[1].type, css.types.STYLE_RULE);
-      assert.equal(tree.rules[2].type, css.types.KEYFRAMES_RULE);
-      assert.equal(tree.rules[3].type, css.types.MEDIA_RULE);
-      assert.equal(tree.rules[3].rules[0].type, css.types.STYLE_RULE);
+      assert.equal(tree.rules[0].type, window.CssParse.types.STYLE_RULE);
+      assert.equal(tree.rules[1].type, window.CssParse.types.STYLE_RULE);
+      assert.equal(tree.rules[2].type, window.CssParse.types.KEYFRAMES_RULE);
+      assert.equal(tree.rules[3].type, window.CssParse.types.MEDIA_RULE);
+      assert.equal(tree.rules[3].rules[0].type, window.CssParse.types.STYLE_RULE);
 
     });
 
     test('rules stringify', function() {
       var orig = sanitizeCss(s.textContent);
-      var result = sanitizeCss(css.stringify(tree));
+      var result = sanitizeCss(window.CssParse.stringify(tree));
       assert.equal(result, orig, 'unexpected stringified output');
     });
 
     test('parse correctly ignores @import and comments', function() {
       var s2 = document.querySelector('#test-ignore');
-      var t = css.parse(s2.textContent);
+      var t = window.CssParse.parse(s2.textContent);
       assert.equal(t.rules[0].selector, '.stuff', 'unexpected rule selector');
       assert.equal(t.rules[0].cssText, 'background: red;', 'unexpected rule cssText');
-      var result = sanitizeCss(css.stringify(t));
+      var result = sanitizeCss(window.CssParse.stringify(t));
       assert.equal(result, '.stuff { background: red; }', 'unexpected stringified output');
     });
 
     test('short escape sequences', function() {
       var s3 = document.querySelector('#short-escape-sequence');
-      var t = css.parse(s3.textContent);
+      var t = window.CssParse.parse(s3.textContent);
       assert.equal(t.rules[0].selector, '.\\000033d-model');
       assert.equal(t.rules[1].selector, '.\\000a33d-model');
       assert.equal(t.rules[2].selector, '.\\00b333d-model');
@@ -169,7 +168,7 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
 
     test('multiple consequent spaces in CSS selector', function() {
       var s4 = document.querySelector('#multiple-spaces');
-      var t = css.parse(s4.textContent);
+      var t = window.CssParse.parse(s4.textContent);
       assert.equal(t.rules[0].selector, '.foo .bar');
       assert.equal(t.rules[1].selector, '.foo .bar');
       assert.equal(t.rules[2].selector, '.foo .bar');
