+++ tests/custom-style.html
@@ -1,0 +1,84 @@
+<!doctype html>
+<!--
+@license
+Copyright (c) 2014 The Polymer Project Authors. All rights reserved.
+This code may only be used under the BSD style license found at http://polymer.github.io/LICENSE.txt
+The complete set of authors may be found at http://polymer.github.io/AUTHORS.txt
+The complete set of contributors may be found at http://polymer.github.io/CONTRIBUTORS.txt
+Code distributed by Google as part of the polymer project is also
+subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
+-->
+<script src="../../web-component-tester/browser.js"></script>
+<script src="../../webcomponentsjs/webcomponents-lite.js"></script>
+<script src="module/generated/make-element.js"></script>
+
+<custom-style>
+  <style>
+  html {
+    --foo: {
+      background-color: rgb(123, 123, 123);
+    };
+  }
+  </style>
+</custom-style>
+
+<template id="x-inner">
+  <style>
+    :host {
+      display: block;
+      height: 100px;
+      width: 100px;
+      @apply --foo;
+      @apply --late;
+    }
+  </style>
+</template>
+
+<template id="x-outer">
+  <style>
+  :host {
+    display: block;
+    @apply --late;
+  }
+  </style>
+  <x-inner></x-inner>
+</template>
+
+<x-outer id="target"></x-outer>
+
+<template id="late">
+  <custom-style>
+    <style>
+    html {
+      --late: {
+        border: 2px solid red;
+      };
+    }
+    </style>
+  </custom-style>
+</template>
+
+<script>
+  suite('Custom Style upgrades', function() {
+    suiteSetup(function() {
+      makeElement('x-inner');
+      makeElement('x-outer');
+    });
+    test('custom-style applies to deeply nested elements', function() {
+      var target = document.querySelector('#target');
+      var inner = target.shadowRoot.querySelector('x-inner');
+      assert.equal(getComputedStyle(inner).backgroundColor, 'rgb(123, 123, 123)');
+    });
+    test('late custom-style forces a tree recalc', function(done) {
+      var target = document.querySelector('#target');
+      var inner = target.shadowRoot.querySelector('x-inner');
+      var late = document.querySelector('template#late');
+      document.body.appendChild(document.importNode(late.content, true));
+      setTimeout(function() {
+        assert.equal(getComputedStyle(target).borderTopWidth.trim(), '2px');
+        assert.equal(getComputedStyle(inner).borderTopWidth.trim(), '2px');
+        done();
+      }, 1000);
+    });
+  });
+</script>
