+++ tests/mixin-ordering.html
@@ -1,17 +1,7 @@
 <!doctype html>
 <head>
   <script src="../../web-component-tester/browser.js"></script>
-  <script>
-    if (window.customElements) {
-      document.write('<script src="../../custom-elements/src/native-shim.js"></s' + 'cript>');
-    } else {
-      document.write('<script src="../../custom-elements/custom-elements.min.js"></s' + 'cript>');
-    }
-  </script>
-  <script src="../../webcomponentsjs/src/Template/Template.js"></script>
-  <script src="../../webcomponentsjs/src/WebComponents/lang.js"></script>
-  <script src="../../shadydom/shadydom.min.js"></script>
-  <script src="../../shadycss/shadycss.min.js"></script>
+  <script src="../../webcomponentsjs/webcomponents-lite.js"></script>
   <script src="module/generated/make-element.js"></script>
 </head>
 <body>
