+++ tests/runner.html
@@ -19,7 +19,8 @@
     'apply-shim.html',
     'scoping.html',
     'mixin-ordering.html',
-    'svg.html'
+    'svg.html',
+    'custom-style.html'
   ];
   WCT.loadSuites(suites);
 </script>
