+++ tests/style-transformer.html
@@ -9,7 +9,6 @@ Code distributed by Google as part of the polymer project is also
 subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
 -->
 <script src="../../web-component-tester/browser.js"></script>
-<script src="../../custom-elements/custom-elements.min.js"></script>
 <script>
 window.ShadyCSS = {shimshadow: true};
 </script>
