+++ tests/svg.html
@@ -9,11 +9,6 @@
     window.ShadyCSS = {shimshadow: true};
   </script>
   <script src="../../webcomponentsjs/webcomponents-lite.js"></script>
-  <script>
-    if (window.customElements) {
-      document.write('<script src="../../custom-elements/src/native-shim.js"></s' + 'cript>');
-    }
-  </script>
   <script src="module/generated/svg-in-shadow.js"></script>
 </head>
 <body>
